refactor: Moved code_gen/*.rs to ../
