+++ lib/shiika_ast/src/lib.rs
@@ -1,6 +1,6 @@
 mod token;
 pub use crate::token::Token;
-use shiika_core::{ty, names::*};
+use shiika_core::{names::*, ty};
 
 #[derive(Debug, PartialEq)]
 pub struct Program {
@@ -58,17 +58,20 @@ pub struct AstMethodSignature {
 impl AstMethodSignature {
     // FIXME: this should be in skc_ast2hir but used by skc_corelib
     pub fn typarams(&self) -> Vec<ty::TyParam> {
-        self.typarams.iter().map(|param| {
-            let v = match &param.variance {
-                AstVariance::Invariant => ty::Variance::Invariant,
-                AstVariance::Covariant => ty::Variance::Covariant,
-                AstVariance::Contravariant => ty::Variance::Contravariant,
-            };
-            ty::TyParam {
-                name: param.name.clone(),
-                variance: v,
-            }
-        }).collect::<Vec<_>>()
+        self.typarams
+            .iter()
+            .map(|param| {
+                let v = match &param.variance {
+                    AstVariance::Invariant => ty::Variance::Invariant,
+                    AstVariance::Covariant => ty::Variance::Covariant,
+                    AstVariance::Contravariant => ty::Variance::Contravariant,
+                };
+                ty::TyParam {
+                    name: param.name.clone(),
+                    variance: v,
+                }
+            })
+            .collect::<Vec<_>>()
     }
 }
