+++ lib/shiika_core/src/names.rs
@@ -395,4 +395,3 @@ pub fn resolved_const_name(namespace: Namespace, names: Vec<String>) -> Resolved
 pub fn typaram_as_resolved_const_name(name: impl Into<String>) -> ResolvedConstName {
     resolved_const_name(Namespace::root(), vec![name.into()])
 }
-
