+++ lib/shiika_core/src/ty.rs
@@ -461,4 +461,3 @@ impl TyParam {
         }
     }
 }
-
